add Polygon Requirements
We are targeting the LATAM public goods - Refi project(s) track, we created `Virtual Study Group App` and app where students from Latin America come together to prepare for web3 hackathons and learn about crypto along with other students that are studying for the same subject matter. The app can provide them with study material, tools, discussion helps, guides, and financial support. This business idea presents a range of challenges but as a business grows we are planning on reaching new cities and scaling globally.